requirements.txt
gspread==5.12.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
pandas==2.1.4
openpyxl==3.1.2
psycopg2-binary==2.9.9
flask==3.0.0